extract existing code to avoid circular dependencies
